+++ examples/modlist.rb
@@ -13,7 +13,7 @@ TT.run(EMAIL, PASSWORD, :room => ROOM) do
   on :user_spoke do |message|
     # Response to "/mod" command
     if moderator_ids.include?(message.sender.id) && message.content =~ /^\/mod$/
-      user.say("Yo #{message.sender.name}, it looks like you are a bot moderator!")
+      room.say("Yo #{message.sender.name}, it looks like you are a bot moderator!")
     end
   end
 end